move object types logic to add_important_values filter
