+++ packages/jira-adapter/src/filters/add_important_values.ts
@@ -5,7 +5,19 @@
  *
  * CERTAIN THIRD PARTY SOFTWARE MAY BE CONTAINED IN PORTIONS OF THE SOFTWARE. See NOTICE FILE AT https://github.com/salto-io/salto/blob/main/NOTICES
  */
-import { CORE_ANNOTATIONS, Element, isObjectType } from '@salto-io/adapter-api'
+import {
+  BuiltinTypes,
+  CORE_ANNOTATIONS,
+  Element,
+  Field,
+  InstanceElement,
+  isInstanceElement,
+  isObjectType,
+  ListType,
+  ObjectType,
+} from '@salto-io/adapter-api'
+import _ from 'lodash'
+import { collections } from '@salto-io/lowerdash'
 import { ImportantValues } from '@salto-io/adapter-utils'
 import { FilterCreator } from '../filter'
 import {
@@ -24,6 +36,9 @@ import {
   WEBHOOK_TYPE,
 } from '../constants'
 import { FIELD_TYPE_NAME } from './fields/constants'
+import { PROJECT_SCOPE_FIELD_NAME } from './projects_scope'
+
+const { makeArray } = collections.array
 
 const importantValuesMap: Record<string, ImportantValues> = {
   [APPLICATION_PROPERTY_TYPE]: [{ value: 'type', highlighted: false, indexed: true }],
@@ -95,6 +110,34 @@ const importantValuesMap: Record<string, ImportantValues> = {
   [WEBHOOK_TYPE]: [{ value: 'Enabled', highlighted: false, indexed: true }],
 }
 
+const createProjectScopeField = (objectType: ObjectType): Field =>
+  new Field(objectType, PROJECT_SCOPE_FIELD_NAME, new ListType(BuiltinTypes.STRING), {
+    [CORE_ANNOTATIONS.HIDDEN_VALUE]: true,
+  })
+
+const addProjectScopeAnnotation = (objectType: ObjectType): void => {
+  objectType.annotations[CORE_ANNOTATIONS.IMPORTANT_VALUES] = [
+    ...makeArray(objectType.annotations[CORE_ANNOTATIONS.IMPORTANT_VALUES]),
+    {
+      value: PROJECT_SCOPE_FIELD_NAME,
+      highlighted: false,
+      indexed: true,
+    },
+  ]
+}
+
+const addProjectScopeToObjectTypes = (instances: InstanceElement[]): void => {
+  const objectTypes = _.uniqBy(
+    instances.map(instance => instance.getTypeSync()),
+    objectType => objectType.elemID.getFullName(),
+  )
+
+  objectTypes.forEach(objectType => {
+    objectType.fields[PROJECT_SCOPE_FIELD_NAME] = createProjectScopeField(objectType)
+    addProjectScopeAnnotation(objectType)
+  })
+}
+
 // Adds relevant important values for the Jira adapter
 const filter: FilterCreator = () => ({
   name: 'addImportantValues',
@@ -107,6 +150,8 @@ const filter: FilterCreator = () => ({
         obj.annotations[CORE_ANNOTATIONS.IMPORTANT_VALUES] = importantValuesArray
       }
     })
+    // projectsScope field
+    addProjectScopeToObjectTypes(elements.filter(isInstanceElement))
   },
 })
