+++ packages/jira-adapter/test/filters/projects_scope.test.ts
@@ -73,73 +73,6 @@ describe('projectsScopeFilter', () => {
     })
   })
 
-  describe('object types', () => {
-    let objectType1: ObjectType
-    let objectType2: ObjectType
-    let objectType3: ObjectType
-    let instance1: InstanceElement
-    let instance2: InstanceElement
-    let instance3: InstanceElement
-
-    beforeEach(async () => {
-      objectType1 = createEmptyType('objectType1')
-      objectType2 = createEmptyType('objectType2')
-      objectType3 = createEmptyType('objectType3')
-      objectType2.annotations[CORE_ANNOTATIONS.IMPORTANT_VALUES] = []
-      objectType3.annotations[CORE_ANNOTATIONS.IMPORTANT_VALUES] = [
-        { value: 'quack', highlighted: false, indexed: true },
-      ]
-      instance1 = new InstanceElement('instance1', objectType1)
-      instance2 = new InstanceElement('instance2', objectType2)
-      instance3 = new InstanceElement('instance3', objectType3)
-    })
-
-    it('should add projectsScope as a hidden field to object types', async () => {
-      const objectTypes = [objectType1, objectType2, objectType3]
-      await filter.onFetch([instance1, instance2, instance3, ...objectTypes])
-
-      objectTypes.forEach(objectType => {
-        expect(objectType.fields.projectsScope).toBeDefined()
-        expect(objectType.fields.projectsScope.annotations[CORE_ANNOTATIONS.HIDDEN_VALUE]).toBeTrue()
-      })
-    })
-
-    it('should add projectsScope to important values annotation', async () => {
-      const objectTypes = [objectType1, objectType2, objectType3]
-      await filter.onFetch([instance1, instance2, instance3, ...objectTypes])
-      expect(objectType1.annotations[CORE_ANNOTATIONS.IMPORTANT_VALUES]).toEqual([
-        { value: 'projectsScope', highlighted: false, indexed: true },
-      ])
-      expect(objectType2.annotations[CORE_ANNOTATIONS.IMPORTANT_VALUES]).toEqual([
-        { value: 'projectsScope', highlighted: false, indexed: true },
-      ])
-      expect(objectType3.annotations[CORE_ANNOTATIONS.IMPORTANT_VALUES]).toEqual([
-        { value: 'quack', highlighted: false, indexed: true },
-        { value: 'projectsScope', highlighted: false, indexed: true },
-      ])
-    })
-
-    it('should not add projectsScope to important values annotation if it already exists', async () => {
-      objectType1.annotations[CORE_ANNOTATIONS.IMPORTANT_VALUES] = [
-        { value: 'projectsScope', highlighted: false, indexed: true },
-      ]
-      await filter.onFetch([instance1, objectType1])
-      expect(objectType1.annotations[CORE_ANNOTATIONS.IMPORTANT_VALUES]).toEqual([
-        { value: 'projectsScope', highlighted: false, indexed: true },
-      ])
-    })
-
-    it('should not add projectsScope important values annotation to object types that does not represent instances', async () => {
-      await filter.onFetch([objectType1])
-      expect(objectType1.annotations[CORE_ANNOTATIONS.IMPORTANT_VALUES]).toBeUndefined()
-    })
-
-    it('should not add projectsScope field to object types that does not represent instances', async () => {
-      await filter.onFetch([objectType1])
-      expect(objectType1.fields.projectsScope).toBeUndefined()
-    })
-  })
-
   describe('references from projects', () => {
     let project1Ref11: InstanceElement
     let project1Ref12: InstanceElement
